Create Resistance-Level-Detect
The function is desinged to detect and build the resistance levels for the stock price chart, so it is an automated simple technical analysis for stock-trading bots. It requires 5 inputs, which are:

HPvec - the list of the "highs" (highest ponts of the stock price chart). Intialized as an empty list in the main body of the algorithm and updated inside the function. 
IndexVec - the list of the position indices of the price candles of the "highs" (if the first candle in the day has index=1 and if the "high" was reached on the 5th candle of the day, then number 5 is added to the list). Initialized as an empty list in the main body of the algorithm and updated inside the function.
highc - the list of the values of the highest points (highs) of all candles (each candle has 5 main parameters (values): high, low, open, close, volume). Initialized as an empty list in the main body of the algorithm and updated outside the function.
current_high - the value of the last "high" which was detected. Initialized as a "0" in the main body of the algorithm and is updated inside the function.
ResElVec - the list of the detected levels (each level is a small list, which consists of two values: the highest border and the lowest border of the level zone). Intialized as an empty list in the main body of the algorithm and updated inside the function.