Create Support-Level-Detect
The function is desinged to detect and build the support levels for the stock price chart, so it is an automated simple technical analysis for stock-trading bots. It requires 5 inputs, which are:

LPvec - the list of the "lows" (lowest ponts of the stock price chart). Intialized as an empty list in the main body of the algorithm and updated inside the function. 
IndexVecL - the list of the position indices of the price candles of the "lows" (if the first candle in the day has index=1 and if the "low" was reached on the 5th candle of the day, then number 5 is added to the list). Initialized as an empty list in the main body of the algorithm and updated inside the function.
lowc - the list of the values of the lowest points (lows) of all candles (each candle has 5 main parameters (values): high, low, open, close, volume). Initialized as an empty list in the main body of the algorithm and updated outside the function.
current_low - the value of the last "low" which was detected. Initialized as a "0" in the main body of the algorithm and is updated inside the function.
SupElVec - the list of the detected levels (each level is a small list, which consists of two values: the highest border and the lowest border of the level zone). Intialized as an empty list in the main body of the algorithm and updated inside the function.